Поправил варнинг

+++ ogsr_engine/xr_3da/SkeletonMotions.h
@@ -185,7 +185,7 @@ struct 	ENGINE_API	motions_value
 	MotionVec*			bone_motions		(shared_str bone_name);
 
 	u32					mem_usage			(){ 
-		u32 sz			=	sizeof(*this)+m_motion_map.size()*6+m_partition.mem_usage();
+		u32 sz			=	u32( sizeof(*this)+m_motion_map.size()*6+m_partition.mem_usage() );
         for (MotionDefVecIt it=m_mdefs.begin(); it!=m_mdefs.end(); it++)
 			sz			+=	it->mem_usage();
 		for (BoneMotionMapIt bm_it=m_motions.begin(); bm_it!=m_motions.end(); bm_it++)
